Did proper session management.
USSD sessions terminate after 20 or so seconds of inactivity. This
would be a pain so we restore session if it hasn't been too long
since previous session. This decision to restore is presented to
the user for them to decide. Start Menu never presents this prompt
to restore as this would be daft. Likewise the Restore Menu never
tries to restore itself.

Updated the parsing to be cleaning so views look much better.

TODO:
  * Make jabber work through django so we don't have to different
    execution paths.
  * Look to implementing notifications on purchase.
  * Look at that django cart example to see if I can drop it in
    quickly.