Finished checkout, and order management.
Now I just need to figure out why delivery costs are not appearing
and determine if what I've done in modifiers and backend is why it
is not working.

So basically on spaza user side still needs:
  Shipping Costs
  COD (Payment backend for the moment)

Things to think about in future would be delivery confirmation, so
taxi gets money, taxi interface to figure out where to deliver and
check confirmation has been received. Add email field to wholesaler
so we can contact them with order. Format order for wholesaler so
they see what they need to pack and can confirm they have shipped
it off with a given taxi.